From Luciano: Hi, here's a new diff ( not for 3.2 release ). It fixes many bugs and adds 2 new mini-editors to the css editor: .. a combo with predefined HTML colors .. a font family chooser dialog
As for the css selector there's a new functionality : based on the
document DTD it shows only the valid tags for this DTD. It is
experimental and for css files you must choose the DTD you want to use (
at the moment the only correct DTD supported is MathML ). The @rules tab
has been partially disabled as I decided to remove it and make an new
editor for  @rules.

The diff doesn't fix css description.rc problem and slotInsertCSS problem.