Another big patch from Jens + some fixes and optimizations by me. Bump version number.
"
Here is the result of my efforts to clean up. I changed the code so that the
build in actions of the kate part are used instead of code in Quanta.

This patch does:

- enable the actions from kate part, with a few exceptions that need some more
attention or are just not wanted in Quanta

- remove a lot of code :-)

- change quantaui.rc so that the merge from the GUI of the kata part and
Quanta gives good results

- add a partmanager, so that the menus and toolbars are changed according to
the active part

- fix a small problem in the project treeview (my last patch for this was not
in CVS so it is in here again)"