quanta.sourceforge.net=>quanta.kdewebdev.org
Unfortunately I couldn't change in every place, due to the string freeze, so we need to redirect the old site when the new one is up (or vice versa).

Also don't be confused if the there is a space between the mimetypes in the list in the description.rc

Update version number.

CCMAIL: sequitur@kde.org