New toolbar behavior in VPL - toggable actions. It's only available with KDE >= 3.4.0 as it needs "KAction::activated(KAction::ActivationReason reason, Qt::ButtonState state)".
Some actions, e.g. "bold", should be toggable, like in any modern editor.
The actions are defined to be toggable in the file "standard.actions", located inside the tarball "quanta/data/toolbars/html/standard.toolbar.tgz", by setting the attribute "toggable" to "true, e.g.:

<action shortcut="Alt+Ctrl+B" icon="tag_bold" type="tag" tooltip="Bold &amp;lt;em>" name="tag_b" text="Bold" toggable="true" >

An action has a scope. For example, if one presses the bold action in the middle of a word, the action should applied in that word, i.e., the scope is "word" in this situation. If the action would be an horizontal alignment change, the scope would be an entire paragraph. The scope of the tag is defined in the "*.tag" files inside "quanta/dtep/", e.g.:

<tag name="strong" hasCore="1" hasI18n="1" hasScript="1" scope="word">

Only "bold" action was changed to behave like this, yet ("standard.actions" and "strong.tag" files). Also, actions with different scope than "word" are not yet implemented.

CCBUG: 99340
CCMAIL: quanta@mail.kde.org, quanta-devel@kde.org