Adjust to use the cmake variables pointing to the platform libraries.
Also kdevplatformeditor was removed recently, but apparently some people are
not yet aware of the policy to build kdewebdev/quanta when doing such changes
to the platform. 

So here is again a reminder to all of us that we need to build kdewebdev/quanta
before checking in changes to kdevplatform that break BC or SC and fix the
quanta build.
CCMAIL:kdevelop-devel@kdevelop.org